Extract shared diagnostic helpers into DiagnosticUtil
